Refactor: move functional tests into test folder, updated test script in package.json
