+++ server/api/users.spec.js
@@ -1,32 +1,32 @@
-/* global describe beforeEach it */
+// /* global describe beforeEach it */
 
-const {expect} = require('chai')
-const request = require('supertest')
-const db = require('../db')
-const app = require('../index')
-const User = db.model('user')
+// const {expect} = require('chai')
+// const request = require('supertest')
+// const db = require('../db')
+// const app = require('../index')
+// const User = db.model('user')
 
-describe('User routes', () => {
-  beforeEach(() => {
-    return db.sync({force: true})
-  })
+// describe('User routes', () => {
+//   beforeEach(() => {
+//     return db.sync({force: true})
+//   })
 
-  describe('/api/users/', () => {
-    const codysEmail = 'cody@puppybook.com'
+//   describe('/api/users/', () => {
+//     const testEmail = 'testemail@test.com'
+//     const testPassword = 'password'
 
-    beforeEach(() => {
-      return User.create({
-        email: codysEmail
-      })
-    })
+//     beforeEach(() => {
+//       return User.create({
+//         testEmail,
+//         testPassword,
+//       })
+//     })
 
-    it('GET /api/users', async () => {
-      const res = await request(app)
-        .get('/api/users')
-        .expect(200)
+//     it('GET /api/users', async () => {
+//       const res = await request(app).get('/api/users').expect(200)
 
-      expect(res.body).to.be.an('array')
-      expect(res.body[0].email).to.be.equal(codysEmail)
-    })
-  }) // end describe('/api/users')
-}) // end describe('User routes')
+//       expect(res.body).to.be.an('array')
+//       expect(res.body[0].email).to.be.equal(testEmail)
+//     })
+//   }) // end describe('/api/users')
+// }) // end describe('User routes')
